requirements.txt
# Medicare FAQ 语义搜索系统依赖
requests>=2.25.0
datasets>=2.0.0
torch>=1.9.0
sentence-transformers>=2.2.0
huggingface_hub>=0.10.0
feedparser>=6.0.0
bs4>=0.0.1
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0